feat: register product

+++ src/lib/firebase/storage.ts
@@ -4,12 +4,31 @@ export const uploadFile = async (
   path: string,
   file: Blob | Uint8Array | ArrayBuffer,
 ) => {
-  const storageRef = storage().ref(path);
-  const snapshot = await storageRef.put(file);
-  return snapshot.ref.getDownloadURL();
+  if (file) {
+    const storageRef = storage().ref(path);
+
+    if (storageRef) {
+      await storageRef.put(file);
+      return storageRef.getDownloadURL();
+    }
+    return null;
+  }
+  console.log('Arquivo não encontrado localmente');
+  return null;
 };
 
 export const deleteFile = async (path: string) => {
   const storageRef = storage().ref(path);
-  await storageRef.delete();
+
+  const fileExists = await storageRef
+    .getDownloadURL()
+    .then(() => true)
+    .catch(() => false);
+
+  if (fileExists) {
+    await storageRef.delete();
+    console.log('Arquivo deletado com sucesso');
+  } else {
+    console.log('Arquivo não encontrado no caminho especificado');
+  }
 };
